v 1.1.0.6
This version implements vote to ban/kick/close.
By default it is not enabled, but can be enabled within the config.ini (EnableVoting)
A moderator or more priveliged user may also enable it from within the room.

Since this have not been tested, i sudgest that DebugToFile is set to True and the DebugLevel is set to 10.

Some information regarding voting.
1) A user MUST have been in the room for atleast 5 minutes before that user can start a vote session.
2) The same applies to voting, a user MUST have been in the room for atleast 5 minutes to vote.
3) A desicion to ban/kick/close will ONLY(this might change) hapend once the vote session is done,
the default time for a vote session is 60 seconds.
4) 1/3 of the room users MUST have voted, when the vote session ends before a desicion will be made.
5) A yes vote should be one of the following '1', 'true', 'True', 'YES', 'Yes', 'yes', 'y', 'Y'.
Likewise, a no vote should be one of the following '0', 'false', 'False', 'NO', 'No', 'no', 'n', 'N'.
Anything else, including a blank vote, will be ignored.
6) Once a user has voted successfully, then a PM will be sent to that user, confirming the vote.

Again THIS HAS NOT BEEN TESTED, so there might be bugs!